修改亮度显示进度0.1%

+++ LY1200_master/LY1200_master/LY1200_APP_master/TouchGFX/gui/include/gui/model/Model.hpp
@@ -22,7 +22,7 @@ public:
 		{
 			return CCT_Temperature_count;
 		}
-		int  getCCTLight()
+		float  getCCTLight()
 		{
 			return CCT_Light_count;
 		}
