+++ LY1200_master/LY1200_master/LY1200_APP_master/TouchGFX/gui/include/gui/screen1_screen/Screen1View.hpp
@@ -17,7 +17,7 @@ public:
 		virtual void TemperatureUp();
 		virtual void TemperatureDown();	
 protected:
-	 int Light_count;
+	 float Light_count;
    int Temperature_count;
 };
